+++ sdk/dotnet/PowerBI/Enums.cs
@@ -66,4 +66,72 @@ namespace Pulumi.AzureNative.PowerBI
 
         public override string ToString() => _value;
     }
+
+    /// <summary>
+    /// Status of the connection.
+    /// </summary>
+    [EnumType]
+    public readonly struct PersistedConnectionStatus : IEquatable<PersistedConnectionStatus>
+    {
+        private readonly string _value;
+
+        private PersistedConnectionStatus(string value)
+        {
+            _value = value ?? throw new ArgumentNullException(nameof(value));
+        }
+
+        public static PersistedConnectionStatus Pending { get; } = new PersistedConnectionStatus("Pending");
+        public static PersistedConnectionStatus Approved { get; } = new PersistedConnectionStatus("Approved");
+        public static PersistedConnectionStatus Rejected { get; } = new PersistedConnectionStatus("Rejected");
+        public static PersistedConnectionStatus Disconnected { get; } = new PersistedConnectionStatus("Disconnected");
+
+        public static bool operator ==(PersistedConnectionStatus left, PersistedConnectionStatus right) => left.Equals(right);
+        public static bool operator !=(PersistedConnectionStatus left, PersistedConnectionStatus right) => !left.Equals(right);
+
+        public static explicit operator string(PersistedConnectionStatus value) => value._value;
+
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public override bool Equals(object? obj) => obj is PersistedConnectionStatus other && Equals(other);
+        public bool Equals(PersistedConnectionStatus other) => string.Equals(_value, other._value, StringComparison.Ordinal);
+
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public override int GetHashCode() => _value?.GetHashCode() ?? 0;
+
+        public override string ToString() => _value;
+    }
+
+    /// <summary>
+    /// Provisioning state of the Private Endpoint Connection.
+    /// </summary>
+    [EnumType]
+    public readonly struct ResourceProvisioningState : IEquatable<ResourceProvisioningState>
+    {
+        private readonly string _value;
+
+        private ResourceProvisioningState(string value)
+        {
+            _value = value ?? throw new ArgumentNullException(nameof(value));
+        }
+
+        public static ResourceProvisioningState Creating { get; } = new ResourceProvisioningState("Creating");
+        public static ResourceProvisioningState Updating { get; } = new ResourceProvisioningState("Updating");
+        public static ResourceProvisioningState Deleting { get; } = new ResourceProvisioningState("Deleting");
+        public static ResourceProvisioningState Succeeded { get; } = new ResourceProvisioningState("Succeeded");
+        public static ResourceProvisioningState Canceled { get; } = new ResourceProvisioningState("Canceled");
+        public static ResourceProvisioningState Failed { get; } = new ResourceProvisioningState("Failed");
+
+        public static bool operator ==(ResourceProvisioningState left, ResourceProvisioningState right) => left.Equals(right);
+        public static bool operator !=(ResourceProvisioningState left, ResourceProvisioningState right) => !left.Equals(right);
+
+        public static explicit operator string(ResourceProvisioningState value) => value._value;
+
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public override bool Equals(object? obj) => obj is ResourceProvisioningState other && Equals(other);
+        public bool Equals(ResourceProvisioningState other) => string.Equals(_value, other._value, StringComparison.Ordinal);
+
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public override int GetHashCode() => _value?.GetHashCode() ?? 0;
+
+        public override string ToString() => _value;
+    }
 }