+++ sdk/go/azure/powerbi/init.go
@@ -21,6 +21,10 @@ func (m *module) Version() semver.Version {
 
 func (m *module) Construct(ctx *pulumi.Context, name, typ, urn string) (r pulumi.Resource, err error) {
 	switch typ {
+	case "azure-native:powerbi:PowerBIResource":
+		r = &PowerBIResource{}
+	case "azure-native:powerbi:PrivateEndpointConnection":
+		r = &PrivateEndpointConnection{}
 	case "azure-native:powerbi:WorkspaceCollection":
 		r = &WorkspaceCollection{}
 	default: