+++ sdk/nodejs/types/enums/powerbi/index.ts
@@ -3,9 +3,11 @@
 
 // Export sub-modules:
 import * as v20160129 from "./v20160129";
+import * as v20200601 from "./v20200601";
 
 export {
     v20160129,
+    v20200601,
 };
 
 export const AzureSkuName = {
@@ -25,3 +27,29 @@ export const AzureSkuTier = {
  * SKU tier
  */
 export type AzureSkuTier = (typeof AzureSkuTier)[keyof typeof AzureSkuTier];
+
+export const PersistedConnectionStatus = {
+    Pending: "Pending",
+    Approved: "Approved",
+    Rejected: "Rejected",
+    Disconnected: "Disconnected",
+} as const;
+
+/**
+ * Status of the connection.
+ */
+export type PersistedConnectionStatus = (typeof PersistedConnectionStatus)[keyof typeof PersistedConnectionStatus];
+
+export const ResourceProvisioningState = {
+    Creating: "Creating",
+    Updating: "Updating",
+    Deleting: "Deleting",
+    Succeeded: "Succeeded",
+    Canceled: "Canceled",
+    Failed: "Failed",
+} as const;
+
+/**
+ * Provisioning state of the Private Endpoint Connection.
+ */
+export type ResourceProvisioningState = (typeof ResourceProvisioningState)[keyof typeof ResourceProvisioningState];