+++ sdk/python/pulumi_azure_native/__init__.py
@@ -11958,6 +11958,8 @@ _utilities.register(
   "mod": "powerbi",
   "fqn": "pulumi_azure_native.powerbi",
   "classes": {
+   "azure-native:powerbi:PowerBIResource": "PowerBIResource",
+   "azure-native:powerbi:PrivateEndpointConnection": "PrivateEndpointConnection",
    "azure-native:powerbi:WorkspaceCollection": "WorkspaceCollection"
   }
  },
@@ -11967,6 +11969,15 @@ _utilities.register(
   "fqn": "pulumi_azure_native.powerbi.v20160129",
   "classes": {
    "azure-native:powerbi/v20160129:WorkspaceCollection": "WorkspaceCollection"
+  }
+ },
+ {
+  "pkg": "azure-native",
+  "mod": "powerbi/v20200601",
+  "fqn": "pulumi_azure_native.powerbi.v20200601",
+  "classes": {
+   "azure-native:powerbi/v20200601:PowerBIResource": "PowerBIResource",
+   "azure-native:powerbi/v20200601:PrivateEndpointConnection": "PrivateEndpointConnection"
   }
  },
  {