+++ sdk/python/pulumi_azure_native/powerbi/_enums.py
@@ -7,6 +7,8 @@ from enum import Enum
 __all__ = [
     'AzureSkuName',
     'AzureSkuTier',
+    'PersistedConnectionStatus',
+    'ResourceProvisioningState',
 ]
 
 
@@ -22,3 +24,25 @@ class AzureSkuTier(str, Enum):
     SKU tier
     """
     STANDARD = "Standard"
+
+
+class PersistedConnectionStatus(str, Enum):
+    """
+    Status of the connection.
+    """
+    PENDING = "Pending"
+    APPROVED = "Approved"
+    REJECTED = "Rejected"
+    DISCONNECTED = "Disconnected"
+
+
+class ResourceProvisioningState(str, Enum):
+    """
+    Provisioning state of the Private Endpoint Connection.
+    """
+    CREATING = "Creating"
+    UPDATING = "Updating"
+    DELETING = "Deleting"
+    SUCCEEDED = "Succeeded"
+    CANCELED = "Canceled"
+    FAILED = "Failed"